+++ CHANGES.md
@@ -2,6 +2,11 @@ Changes
 =======
 0.1.2
 
+2013/05/22 pfraze
+
+ - Added transform-link discovery through response headers
+
+
 2013/05/21 pfraze
 
  - Added region close on status 210